multicluster: Carve out multicluster-related functions in a separate file
